Added functionality to DrinkController Class, basically just a renamed DrinkList class.
